Read A_LEVEL.out

A-level information is printed each time a time-dependent boundary condition is specified. The information is directed to output file A_LEVEL.OUT.

Usage

read_alevel(path, dbg = FALSE)

Arguments

path

full path to A_LEVEL.out file

dbg

show debug messages (default: FALSE)

Value

imports A_LEVEL out with tidy column names and saves metainformation in attributes 'meta_general' and 'meta_units'

time

Time, t, at current time-level [T]

sum_r_top

Cumulative potential surface flux [L] (infiltration/evaporation: -/+)

sum_r_root

Cumulative potential transpiration [L]

sum_v_top

Cumulative value of the actual surface flux [L] (infiltration/evaporation: -/+) [-]

sum_v_root

Cumulative value of the actual transpiration [L]

sum_v_bot

Cumulative value of the bottom boundary flux [L] (inflow/outflow: +/-)

h_top

Pressure head at the soil surface [L]

h_root

Mean value of the pressure head in the soil root zone for which Beta(n)>0 [L]

h_bot

Pressure head at the bottom of the soil profile [L]

a_level

A-level number (current variable boundary condition number) [-]

References

https://www.pc-progress.com/Downloads/Pgm_Hydrus1D/HYDRUS1D-4.17.pdf#page=277

Examples

path_alevel <- system.file("extdata/model/test/A_LEVEL.out", package = "kwb.hydrus1d")
alevel <- read_alevel(path = path_alevel)
alevel
#> # A tibble: 81 × 10
#>     time     sum_r_top sum_r_root sum_v_top sum_v_root sum_v_bot    h_top h_root
#>  * <dbl>         <dbl>      <dbl>     <dbl>      <dbl>     <dbl>    <dbl>  <dbl>
#>  1     1  0.25                  0    0.0698          0  -0.00173  -1   e5      0
#>  2     2 -0.2                   0   -0.380           0  -0.00346  -8.35e1      0
#>  3     3 -0.02                  0   -0.200           0  -0.00520  -2.00e3      0
#>  4     4 -0.14                  0   -0.320           0  -0.00693  -1.05e2      0
#>  5     5 -0.13                  0   -0.310           0  -0.00866  -1.25e2      0
#>  6     6 -0.0000000270          0   -0.209           0  -0.0104   -1   e5      0
#>  7     7  0.22                  0   -0.177           0  -0.0121   -1   e5      0
#>  8     8  0.27                  0   -0.161           0  -0.0139   -1   e5      0
#>  9     9  0.45                  0   -0.146           0  -0.0156   -1   e5      0
#> 10    10  0.63                  0   -0.135           0  -0.0173   -1   e5      0
#> # ℹ 71 more rows
#> # ℹ 2 more variables: h_bot <dbl>, a_level <dbl>