Read Run_Inf.out

Contains time and iteration information

Usage

read_runinf(path, dbg = FALSE)

Arguments

path

full path to Run_Inf.out file

dbg

show debug messages (default: FALSE)

Value

imports Run_Inf.out with tidy column names and saves metainformation in attributes 'meta_general' and 'meta_units'

tlevel

Time-level (current time-step number) [-]

time

Time, t, at current time-level [T]

dt

Time step, delta t [T]

itr_w

Number of iterations necessary for solution of the water flow equation [-]

itr_c

Number of iterations necessary for solution of the solute transport equation [-]

it_cum

Cumulative number of iterations [-]

kod_t

Code for the boundary condition at the soil surface

kod_b

Code for the boundary condition at the bottom of the soil profile

converg

Information whether or not the numerical convergence was achieved at the current time-level

peclet

Maximum local Peclet number [-]

courant

Maximum local Courant number [-]

References

https://www.pc-progress.com/Downloads/Pgm_Hydrus1D/HYDRUS1D-4.17.pdf#page=272

Examples

path_runinf <- system.file("extdata/model/test/Run_Inf.out", package = "kwb.hydrus1d")
runinf <- read_runinf(path = path_runinf)
runinf
#> # A tibble: 2,272 × 11
#>    t_level  time      dt itr_w itr_c it_cum kod_t kod_b converg peclet courant
#>  *   <dbl> <dbl>   <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <lgl>    <dbl>   <dbl>
#>  1       1 0.111 0.111       5     1     25    -4    -5 TRUE       0.1   0.18 
#>  2       2 0.148 0.0370      3     1     38    -4    -5 TRUE       0.1   0.075
#>  3       3 0.164 0.0160      2     1     50    -4    -5 TRUE       0.1   0.034
#>  4       4 0.171 0.00695     2     1     62    -4    -5 TRUE       0.1   0.015
#>  5       5 0.180 0.00904     3     1     65    -4    -5 TRUE       0.1   0.009
#>  6       6 0.184 0.00392     2     1     77     4    -5 TRUE       0.1   0.003
#>  7       7 0.189 0.00509     2     1     79     4    -5 TRUE       0.1   0.004
#>  8       8 0.196 0.00662     2     1     81     4    -5 TRUE       0.1   0.006
#>  9       9 0.204 0.00861     2     1     83     4    -5 TRUE       0.1   0.007
#> 10      10 0.216 0.0112      2     1     85     4    -5 TRUE       0.1   0.009
#> # ℹ 2,262 more rows