Read T_LEVEL.out

Stores pressure heads and fluxes on the boundaries and in the root zone.

Usage

read_tlevel(path, dbg = FALSE)

Arguments

path

full path to T_LEVEL.out file

dbg

show debug messages (default: FALSE)

Value

imports T_LEVEL out with tidy column names and saves metainformation in attributes 'meta_general' and 'meta_units'

time

Time, t, at current time-level [T]

r_top

Potential surface flux [LT-1] (infiltration/evaporation: -/+)

r_root

Potential transpiration rate [LT-1]

v_top

Actual surface flux [LT-1] (infiltration/evaporation: -/+)

v_root

Actual transpiration rate [LT-1]

v_bot

Actual flux across the bottom of the soil profile [LT-1] (inflow/outflow +/-)

sum_r_top

Cumulative value of the potential surface flux [L] (infiltration/evaporation: -/+)

sum_r_root

Cumulative value of the potential transpiration rate [L]

sum_v_top

Cumulative value of the actual surface flux [L]

sum_v_root

Cumulative value of the actual transpiration rate [L] (infiltration/evaporation: -/+)

sum_v_bot

Cumulative value of the actual flux across the bottom of the soil profile [LT-1] (inflow/outflow +/-)

h_top

Pressure head at the soil surface [L]

h_root

Mean value of the pressure head over the region for which Beta(n) > 0 (i.e. within the root zone) [L]

h_bot

Pressure head at the bottom of the soil profile [L]

run_off

Surface runoff [LT-1]

sum_run_off

Cumulative surface runoff [L]

volume

Volume of water in the entire flow domain [L]

sum_infil

Cumulative infiltration [L]

sum_evap

Cumulative actual evaporation [L]

t_level

Time-level (current time-step number) [-]

cum_w_trans

Cumulative mass transfer between the mobile and immobile regions for dual porosity model [L]

snow_layer

Thickness of snow layer, expressed as the "snow water equivalent" (the amount of water contained within the snowpack) [L]

References

https://www.pc-progress.com/Downloads/Pgm_Hydrus1D/HYDRUS1D-4.17.pdf#page=271

Examples

path_tlevel <- system.file("extdata/model/test/T_LEVEL.out", package = "kwb.hydrus1d")
tlevel <- read_tlevel(path = path_tlevel)
tlevel
#> # A tibble: 2,272 × 22
#>     time r_top r_root  v_top v_root    v_bot sum_r_top sum_r_root sum_v_top
#>  * <dbl> <dbl>  <dbl>  <dbl>  <dbl>    <dbl>     <dbl>      <dbl>     <dbl>
#>  1 0.111  0.25      0 0.25        0 -0.00173    0.0278          0    0.0278
#>  2 0.148  0.25      0 0.25        0 -0.00173    0.0370          0    0.0370
#>  3 0.164  0.25      0 0.25        0 -0.00173    0.0410          0    0.0410
#>  4 0.171  0.25      0 0.25        0 -0.00173    0.0428          0    0.0428
#>  5 0.180  0.25      0 0.0977      0 -0.00173    0.0450          0    0.0437
#>  6 0.184  0.25      0 0.0786      0 -0.00173    0.0460          0    0.0440
#>  7 0.189  0.25      0 0.0819      0 -0.00173    0.0473          0    0.0444
#>  8 0.196  0.25      0 0.0840      0 -0.00173    0.0490          0    0.0450
#>  9 0.204  0.25      0 0.0816      0 -0.00173    0.0511          0    0.0457
#> 10 0.216  0.25      0 0.0774      0 -0.00173    0.0539          0    0.0465
#> # ℹ 2,262 more rows
#> # ℹ 13 more variables: sum_v_root <dbl>, sum_v_bot <dbl>, h_top <dbl>,
#> #   h_root <dbl>, h_bot <dbl>, run_off <dbl>, sum_run_off <dbl>, volume <dbl>,
#> #   sum_infil <dbl>, sum_evap <dbl>, t_level <dbl>, cum_w_trans <dbl>,
#> #   snow_layer <dbl>