Read solute1.out

Reads Solute output

Usage

read_solute(
  path = system.file("extdata/model/test/solute1.out", package = "kwb.hydrus1d"),
  dbg = FALSE
)

Arguments

path

full path to solute_id.out file (default: system.file("extdata/model/test/solute1.out", package = "kwb.hydrus1d"))

dbg

show debug messages (default: FALSE)

Value

imports solute_id.out with tidy column names and saves metainformation in attributes 'meta_general' and 'meta_units'

time

Time, t, at current time-level [T]

cv_top

Actual solute flux across the soil surface [ML-2 T -1] (inflow/outflow: +/-)

cv_bot

Actual solute flux across the bottom of the soil profile [ML-2 T -1] (inflow/outflow: +/-)

sum_cv_top

Cumulative solute flux across the soil surface [ML-2] (inflow/outflow: +/-)

sum_cv_bot

Cumulative solute flux across the bottom of the soil profile [ML-2] (inflow/outflow: +/-)

sum_cv_ch0

Cumulative amount of solute added to the flow region by zero-order reactions [ML-2] (negative when removed from the system)

sum_cv_ch1

Cumulative amount of solute removed from the flow region by first-order reactions [ML-2] (negative removed from the system).

c_top

Solute concentration at the soil surface [ML-3]

c_root

Mean solute concentration of the root zone [ML-3]

c_bot

Solute concentration at the bottom of the soil profile [ML-3]

cv_root

Actual root solute uptake in the root zone [ML-2 T -1] (positive when removed from the system)

sum_cv_root

Cumulative amount of solute removed from the flow region by root water uptake S [ML-2] (positive when removed from the system)

sum_cv_n_eql

Cumulative mass transfer to either kinetic adsorption sites (type-2 adsorption sites), or to the immobile liquid region [ML-2] (inflow/outflow: +/-).

t_level

Time-level (current time-step number) [-]

c_gwl

Average concentration in the saturated zone [ML-3] (in the groundwater)

c_run_off

Solute flux in the runoff ([ML-3] * [LT-1]) [ML-2T-1]

sum_c_run_off

Cumulative solute flux in the runoff ([ML-3] * [LT-1]* [T]) [ML-2]

cv_i

Solute flux at the first through third observation node ([ML-3] * [LT-1]) [ML-2T-1]

sum_cv_i

cumulative solute flux at the first through third observation node ([ML-3] * [LT-1] * [T]) [ML-2]. Total solute fluxes are reported when only one solute is simulated. Only convective fluxes (for the first solute) are reported when 2 or more solutes are simulated

References

https://www.pc-progress.com/Downloads/Pgm_Hydrus1D/HYDRUS1D-4.17.pdf#page=273

Examples

path_solute <- system.file("extdata/model/test/solute1.out", package = "kwb.hydrus1d")
solute <- read_solute(path = path_solute)
solute
#> # A tibble: 2,272 × 19
#>     time cv_top cv_bot sum_cv_top sum_cv_bot cv_ch0 cv_ch1 c_top c_root c_bot
#>  * <dbl>  <dbl>  <dbl>      <dbl>      <dbl>  <dbl>  <dbl> <dbl>  <dbl> <dbl>
#>  1 0.111      0      0          0          0      0      0     0      0     0
#>  2 0.148      0      0          0          0      0      0     0      0     0
#>  3 0.164      0      0          0          0      0      0     0      0     0
#>  4 0.171      0      0          0          0      0      0     0      0     0
#>  5 0.180      0      0          0          0      0      0     0      0     0
#>  6 0.184      0      0          0          0      0      0     0      0     0
#>  7 0.189      0      0          0          0      0      0     0      0     0
#>  8 0.196      0      0          0          0      0      0     0      0     0
#>  9 0.204      0      0          0          0      0      0     0      0     0
#> 10 0.216      0      0          0          0      0      0     0      0     0
#> # … with 2,262 more rows, and 9 more variables: cv_root <dbl>,
#> #   sum_cv_root <dbl>, sum_cv_n_eql <dbl>, t_level <dbl>, c_gwl <dbl>,
#> #   c_run_off <dbl>, sum_c_run_off <dbl>, cv_i <dbl>, sum_cv_i <dbl>