Read T_LEVEL.out

Read T_LEVEL.out

Usage

read_tlevel(path)

Arguments

path

full path to T_LEVEL.out file

Value

imports T_LEVEL out with tidy column names and saves metainformation in attribute 'meta'

Examples

path_tlevel <- system.file("extdata/model/test/T_LEVEL.out", package = "kwb.hydrus1d")
tlevel <- read_tlevel(path = path_tlevel)
#> Rows: 17322 Columns: 22
#> ── Column specification ────────────────────────────────────────────────────────
#> 
#> dbl (22): time, r_top, r_root, v_top, v_root, v_bot, sum_r_top, sum_r_root, ...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
tlevel
#> # A tibble: 17,322 × 22
#>     time r_top r_root  v_top v_root   v_bot sum_r_top sum_r_root sum_v_top
#>    <dbl> <dbl>  <dbl>  <dbl>  <dbl>   <dbl>     <dbl>      <dbl>     <dbl>
#>  1 0.111  0.25      0 0.25        0 -0.0119    0.0278          0    0.0278
#>  2 0.222  0.25      0 0.25        0 -0.0119    0.0556          0    0.0556
#>  3 0.259  0.25      0 0.0936      0 -0.0119    0.0648          0    0.0590
#>  4 0.265  0.25      0 0.113       0 -0.0119    0.0662          0    0.0596
#>  5 0.272  0.25      0 0.0678      0 -0.0119    0.0679          0    0.0601
#>  6 0.281  0.25      0 0.0600      0 -0.0119    0.0702          0    0.0606
#>  7 0.284  0.25      0 0.0892      0 -0.0119    0.0711          0    0.0610
#>  8 0.290  0.25      0 0.101       0 -0.0119    0.0724          0    0.0615
#>  9 0.296  0.25      0 0.0930      0 -0.0119    0.0741          0    0.0621
#> 10 0.305  0.25      0 0.0986      0 -0.0119    0.0762          0    0.0630
#> # … with 17,312 more rows, and 13 more variables: sum_v_root <dbl>,
#> #   sum_v_bot <dbl>, h_top <dbl>, h_root <dbl>, h_bot <dbl>, run_off <dbl>,
#> #   sum_run_off <dbl>, volume <dbl>, sum_infil <dbl>, sum_evap <dbl>,
#> #   t_level <dbl>, cum_w_trans <dbl>, snow_layer <dbl>