Write "ATMOSPH.IN" input file

Write "ATMOSPH.IN" input file

Usage

write_atmosphere(
  atm,
  MaxAL = 365,
  DailyVar = FALSE,
  SinusVar = FALSE,
  lLai = FALSE,
  lBCCycles = FALSE,
  lInterc = FALSE,
  hCritS = 0
)

Arguments

atm

tibble of input data as defined in prepare_atmospherice_input

MaxAL

Number of meteorological records

DailyVar

TRUE if HYDRUS-1D is to generate daily variations in evaporation and transpiration (see section 2.7.2.)., otherwise: FALSE (default: FALSE)

SinusVar

TRUE if HYDRUS-1D is to generate sinusoidal variations in precipitation (see section 2.7.2.), otherwise: FALSE(default: FALSE)

lLai

Logical variable indicating that potential evapotranspiration is to be divided into potential evaporation and potential transpiration using eq. (2.75). (default: FALSE)

lBCCycles

TRUE if a set of boundary conditions is to be repeated multiple times, otherwise FALSE(default: FALSE)

lInterc

TRUE if interception is considered using eq. (2.78), otherwise FALSE (default: FALSE)

hCritS

Maximum allowed pressure head at the soil surface (L). (default: 0)

Value

Creates ATMOSPH.IN input textfile

Examples

write_atmosphere()
#> Error in convert_atmosphere_to_string(atm): argument "atm" is missing, with no default